% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progressbar.r
\docType{methods}
\name{kill}
\alias{kill}
\alias{kill,ProgressBar-method}
\title{Kills current ProgressBar.}
\usage{
kill(x)

\S4method{kill}{ProgressBar}(x)
}
\arguments{
\item{x}{`ProgressBar` instance}
}
\description{
Kills current ProgressBar.
}

