% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progressbar.r
\docType{class}
\name{ProgressBar}
\alias{ProgressBar}
\alias{ProgressBar-class}
\title{ProgressBar with labels}
\arguments{
\item{...}{don't know if used}
}
\description{
Classe S4 per ProgressBar
}
\section{Slots}{

\describe{
\item{\code{value}}{current value}

\item{\code{min}}{minimum value}

\item{\code{max}}{maximum value}

\item{\code{char}}{char to use as token for progress}
}}

