% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/progressbar.r
\docType{class}
\name{ProgressBar}
\alias{ProgressBar}
\alias{ProgressBar-class}
\title{ProgressBar with labels}
\description{
Classe S4 per ProgressBar
}
\section{Slots}{

\describe{
\item{\code{value}}{current value}

\item{\code{min}}{minimum value}

\item{\code{max}}{maximum value}

\item{\code{char}}{char to use as token for progress}

\item{\code{width}}{current screen width, autoevaluated}

\item{\code{time}}{time since the beginning of ProgressBar}
}}

